feat: rewrite app logic to generate calendar files
