+++ src/public/index.html
@@ -15,9 +15,9 @@
         <p class="lead">This page provides an iCalendar (RFC 2445) format of the University of York's week numbers, according to the <a href="https://www.york.ac.uk/about/term-dates/" target="_blank">term dates</a>.</p>
         <p>It is compatible with most calendar software, the program provides three different calendars depending on your circumstance. Simply adding one of the following URL to your Calendar software:</p>
         <ul>
-            <li><b>Undergraduate</b> - <code>https://calendar.chrisnethercott.co.uk/calendar?type=ug</code></li>
-            <li><b>Postgraduate</b> - <code>https://calendar.chrisnethercott.co.uk/calendar?type=pg</code></li>
-            <li><b>Staff</b> - <code>https://calendar.chrisnethercott.co.uk/calendar?type=staff</code></li>
+            <li><b>Undergraduate</b> - <code>https://chriscn.github.io/uoy-week-calendar/calendar/undergraduate.ics</code></li>
+            <li><b>Postgraduate</b> - <code>https://chriscn.github.io/uoy-week-calendar/calendar/postgraduate.ics</code></li>
+            <li><b>Staff</b> - <code>https://chriscn.github.io/uoy-week-calendar/calendar/staff.ics</code></li>
         </ul>
         <p>The program should support up to the academic year 2029/2030; and is easily updatable past that. However, it will only generate the current academic year's week numbers.</p>
         <h2>Format</h2>
